Improve provenance modeling.
An Auction of a lot now causes a Procurement activity, which has constituent parts for
Payment and Acquisition.

Sales records which indicate that a single payment was made for multiple lots are now
represented with a single Procurement and Payment record.

The code is now tracking post sales data so that post-processing can reconcile the URIs
of objects that can be identified uniquely as having participated in multiple auctions.