Add modeling support for buyer and seller name_source data.
When the name source can uniquely identify a known physical copy of a catalog,
this is modeled as the hand-written notes (Linguistic Object) in the catalog
referring to the acquisition in which the buyer/seller takes part. In the case
of unsold transactions, the acquisition in question will be the previous
acquisition leading to the current owner who is offering the object for sale.