Add modeling of object materials.
This asserts a made_of property on human made objects based on the 'object_type' and
'materials' fields. These values are compared against the materials_map service JSON
data. If a match is found, a static Material instance is fetched (which was constructed
from the materials service JSON data), and asserted as the part_of value.