Fix race condition on renaming files in parallel during post-processing.
The post-processing step that converts temporary URIs to UUID URNs had a race
condition when run in parallel and set to rename files using the resulting
UUID.

This manifested in at least one case as a dropping of an artist record in a
case where an object appearing in multiple sales had differing artist records.

The trivial fix is to disallow file renaming with parallel processing.
This will result in significantly longer post-processing time.